refactor(cli): extract report writers into cli/report_writers.py
